freewillai/requirements.txt
aioipfs
eth_account
keras
numpy
onnx
onnxruntime
python-dotenv
skl2onnx
tensorflow
torch
web3
scikit-learn
tqdm
py-solc-x
pillow
polars
torchvision
transformers
pandas
puremagic

fastapi
uvicorn
nest_asyncio
RestrictedPython
dagger-io
